Dự án Chatbot FoxAI/TELEGRAM BOT/requirement.txt
python-telegram-bot
python-dotenv
python-OpenAi
